svelte-custom-element: wrapper class for data validation
